fix(store): `BenchmarkLoadValidators` no longer panics (#4367)
### Context
The `BenchmarkLoadValidators` benchmark in the `state` package panics
because it can't find the validators' information for the specified
heights. The issue is caused by a missing `DBKeyLayout` in the test
database configuration, which leads to the `LoadValidators` call
failing.

### This Changes
- the benchmark now creates a new state `Store` by explicitly specifying
the `DBKeyLayout` in the `StoreOptions`.
- `store` constructor checks if `DBKeyLayout` is set; if not, it sets it
to a default value.

---

#### PR checklist

- [x] Tests written/updated
- ~[ ] Changelog entry added in `.changelog` (we use
[unclog](https://github.com/informalsystems/unclog) to manage our
changelog)~
- ~[] Updated relevant documentation (`docs/` or `spec/`) and code
comments~