feat(e2e): Support latency emulation on 'fast' Docker image (#4348)
On #4408 we removed the dependency on Python to run the latency
emulation script, by replacing it for a shell script with just `tc`
(traffic control) commands. Now we are able to run the script in the
slim Docker image generated with `make fast`.

Note that the base image of `Dockerfile.fast` was changed from
`distroless/static-debian11` to `alpine`, because it is easier to
install `tc` and run scripts from `alpine`. The difference in the image
size is not really significant: 48.5MB in distroless vs 56.7MB in
alpine.

---

#### PR checklist

- [ ] Tests written/updated
- [ ] Changelog entry added in `.changelog` (we use
[unclog](https://github.com/informalsystems/unclog) to manage our
changelog)
- [ ] Updated relevant documentation (`docs/` or `spec/`) and code
comments

---------

Co-authored-by: Andy Nogueira <me@andynogueira.dev>
Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>
(cherry picked from commit fe4faff903fcd0e0f73d91ceefa7d91cb4d6ad21)