chore(state): `v1LegacyLayout` keys no longer use `fmt.Sprintf()`. (#4484)
The `v1LegacyLayout` type uses `fmt.Sprintf()` to create the DB keys.
This PR substitutes `fmt.Sprintf` with string concatenation, which is
3-4 times faster for such a simple string.

In general, we shouldn't use `fmt.Sprintf()` to concatenate strings,
because it's inefficient.
Exceptions could be constructing error messages (but we should use
`fmt.Errorf()` for that) or tests.

---

#### PR checklist

- ~[ ] Tests written/updated~
- [x] Changelog entry added in `.changelog` (we use
[unclog](https://github.com/informalsystems/unclog) to manage our
changelog)
- ~[ ] Updated relevant documentation (`docs/` or `spec/`) and code
comments~

---------

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>