Updated `golangci` to allow importing github.com/cockroachdb/pebble.
The linter was failing with the following error:
`import 'github.com/cockroachdb/pebble' is not allowed from list 'main'`.
even though we are not importing github.com/cockroachdb/pebble in the `main` pkg.

I think this happened because the `depguard` linter configuration treats all
files in the project as belonging to the `main` package due to the `$all` rule
in its configuration. That is all files (except test files) are evaluated with
the "main rules", even if they belong to sub-packages.