fix(mempool): Extend `ErrInvalidTx` with `CheckTxResponse` info (#4550)
When calling the RPC endpoints `broadcast_tx_*`, the mempool returns a
`reqRes` object with a potential error. Before #4040, `reqRes` didn't
have an error. An error on the ABCI CheckTx call would be returned
encoded in the fields Code, Log, etc, of `ResultBroadcastTx`. Any other
internal error of the mempool would be discarded.

The problem is that now when the transaction is invalid,
`broadcast_tx_*` return `ErrTxBroadcast{Source: ErrCheckTxFailed,
ErrReason: err}` where `err` is simply `ErrInvalidTx`, without any other
information. This PR adds all the fields in `ResultBroadcastTx` (Code,
Data, Log, Codespace, and Hash) to `ErrInvalidTx`, so that this
information is available gain to the `broadcast_tx_*` caller.

---

#### PR checklist

- [x] Tests written/updated
- [x] Changelog entry added in `.changelog` (we use
[unclog](https://github.com/informalsystems/unclog) to manage our
changelog)
- [ ] Updated relevant documentation (`docs/` or `spec/`) and code
comments

---------

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>