feat(e2e): Support sending the same tx to  multiple nodes (#4513)
Closes #4430 


This PR adds a new flag `--num-nodes-per-tx N ` where N is the number of
nodes to which the loader is sending the same transactions in an
iteration.

If omitted the transactions are sent in round robin. This was added to
facilitate testing the efficiency of the DOG protocol.

---

#### PR checklist

- [ ] Tests written/updated
- [x] Changelog entry added in `.changelog` (we use
[unclog](https://github.com/informalsystems/unclog) to manage our
changelog)
- [ ] Updated relevant documentation (`docs/` or `spec/`) and code
comments

---------

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>
Co-authored-by: Hernán Vanzetto <15466498+hvanz@users.noreply.github.com>