feat(storage): Add `PrefixDB` to pkg `internal/storage` (#4562)
Partially addresses #4486.

### Context
CometBFT will stop importing
[cometbft-db](https://github.com/cometbft/cometbft-db) as a dependency
to support multiple database backends. Instead, it will use
[pebble](https://github.com/cockroachdb/pebble) by default.

### Changes
This PR adds type `PrefixDB` to `internal/storage` from cometbft-db. 
`PrefixDB` is a database wrapper that implements the`DB`, `Batch`, and
`Iterator` interfaces.

**Note**: this PR makes some changes to the implementation of
`PrefixDB`. Namely:
- `PrefixDB` no longer uses a mutex, because we think that it is
sufficient to rely on the wrapped database's concurrency control
mechanism. That is, concurrent calls to `PrefixDB`'s methods will
already "lock" on the calls to the wrapped DB methods.
- `NewPrefixDB` now returns an error if the given prefix is empty. We
think allowing the creation of a `PrefixDB` with no prefix was an
oversight in the design.
- `prefixDBBatch` methods now want a pointer receiver. 
- `PrefixDB.Compact` now prepends the prefix to `start` and `end` keys
defining the iterator's range. The previous implementation did not do
so, and this was likely an oversight. We believed this behavior
contradicted the definition of a `PrefixDB` as a logical database that
prepends the prefix to all keys before operating on the underlying
database.