fix(state): `TxIndexer` no longer leaks its database's goroutines. (#4642)
### Context
The
[`TxIndexer`](https://github.com/cometbft/cometbft/blob/2b1db1c16bf2db16b81b49fef3581e79679fbed6/state/txindex/indexer.go#L17)
interface defines how to index and search transactions. Its implementers
need to interact with a database, which callers are typically expected
to close when done. However, `TxIndexer` does not provide a `Close`
method. This prevents closing the database used by the transaction
indexer, causing goroutines associated with that database to leak.

Two
[tests](https://github.com/cometbft/cometbft/blob/2b1db1c16bf2db16b81b49fef3581e79679fbed6/internal/inspect/inspect_test.go#L29),
`TestInspectConstructor` and `TestInspectRun`, show this problem
indirectly. These tests check whether an
[`Inspector`](https://github.com/cometbft/cometbft/blob/2b1db1c16bf2db16b81b49fef3581e79679fbed6/internal/inspect/inspect.go#L32)
leaks goroutines. The `Inspector` sets up three databases: a block
database, a state database, and a transaction indexer database. It
closes only the block and state databases because it cannot close the
transaction indexer’s database (because of the missing `Close` method).
Therefore, the transaction indexer database's goroutines leak, and the
two tests above detect that.

In `main`, both tests pass because they use a
[`MemDB`](https://github.com/cometbft/cometbft-db/blob/4cf60c715fe8daccb9dce3b24295575bd461d5d8/memdb.go#L52),
an in-memory store that does not spawn goroutines. However, in the [work
to remove cometbft-db](https://github.com/cometbft/cometbft/pull/4601),
the tests switch to a “real” database that does spawn goroutines.
Without a proper `Close` method in `TxIndexer`, these goroutines remain
active, causing `TestInspectConstructor` and `TestInspectRun` to fail.

### This Change 
This PR adds a `Close` method to the `TxIndexer` interface and updates
the existing implementation to properly close the underlying database.
It also modifies the code to call `TxIndexer.Close` where needed,
ensuring that database resources are released and preventing goroutine
leaks.

---

#### PR checklist

- [x] Tests written/updated
- [x] Changelog entry added in `.changelog` (we use
[unclog](https://github.com/informalsystems/unclog) to manage our
changelog)
- [x] Updated relevant documentation (`docs/` or `spec/`) and code
comments