Fixed panic in `node` pkg `TestNodeStartStop` test.
The test unfolds too quickly for the node's indexer to finish its operations.
Therefore, the goroutine below that calls n.Stop() will close the indexer's
database right when the indexer is trying to write to it at line
state/txindex/indexer_service.go:109
thus causing a panic.  This small sleep allows the indexer to finish its
write operation before the node is shut down.
Unfortunately, there is no graceful shutdown available for the database layer,
i.e., make the database wait for all the active connections before closing.

The test does not fail in `main` because the code in `main` does not close
the indexer's database when shutting down a node (which causes its goroutines
to leak. See #4642).