fix(e2e): Reduce size of generated vote extensions and don't convert to Hex (#4721)
PR #4688 increased the size of vote extensions used in the e2e app which
resulted in e2e tests failing because the messages were too big.

This PR does two things:
- Does not convert the Non replay-protected vote extensions to a Hex
String (as this doubles their size).
- Uses a lower value for the generated VEs because this value will in
practice be doubled (if the configuration size is 256, we will have
`VoteExtension` of size 256 and `NonRpExtension` of size 256).


---

#### PR checklist

- [ ] Tests written/updated
- [ ] Changelog entry added in `.changelog` (we use
[unclog](https://github.com/informalsystems/unclog) to manage our
changelog)
- [ ] Updated relevant documentation (`docs/` or `spec/`) and code
comments