docs: document max values for PBTS Precision and MessageDelay (#5046)
Closes #4893 
This PR updates the PBTS-related documentation to reflect the
implementation constraints introduced in #4816.

- In `proposer-based-timestamps.md`, added notes under `Precision` and
`MessageDelay` to document the hard-coded caps:
  - `Precision` ≤ 30s
  - `MessageDelay` ≤ 24h

- In `data_structures.md`, added a note below the SynchronyParams table
to mention these limits as implementation constraints.

- In `pbts-sysmodel.md`, included a note after [PBTS-MSG-DELAY.0]
explaining that although these parameters are bounded in the
implementation, it has no impact on the formal model.

These updates ensure that readers are aware of the practical limitations
applied in the implementation and help prevent misconfiguration or
misunderstanding.
#### PR checklist

- [ ] Tests written/updated
- [ ] Changelog entry added in `.changelog` (we use
[unclog](https://github.com/informalsystems/unclog) to manage our
changelog)
- [X] Updated relevant documentation (`docs/` or `spec/`) and code
comments