fix: update nodekey.ID to use type alias instead of declaration (#5157)
---

Update `nodekey.ID` to be `type ID = string` instead of `type ID
string`.

For some reason mockery attempts to use the internal path when
regenerating mocks when the modules path is update to
`github.com/cometbft/cometbft/v2`.

This change make mockery use string instead of the internal type. I am
unsure why it uses the public type in the v1 implementation.

Long term solution is likely to update mockery, but this fixes it for
now at the cost of reduction in type strictness.

#### PR checklist

- [x] Tests written/updated
- [x] Changelog entry added in `.changelog` (we use
[unclog](https://github.com/informalsystems/unclog) to manage our
changelog)
- [x] Updated relevant documentation (`docs/` or `spec/`) and code
comments