refactor: e2e-node docker build (#5176)
---

The way the old docker file worked, it would build images and tag them,
but that tag gets overwritten by the multi-architecture build.

This is very confusing because there is times during build where the
multi-architecture build only has amd64 for a minute or so, then it only
has arm64 for a minute or so, and assuming the job doesn't fail, it
eventually becomes a true multi architecture image.

This PR instead creates a separate tag for the single architecture build
to make the intermediate steps a bit clearer, and in case of failure
doesn't accidently leave a multi-arch build in a single-arch state.

This can safely be backported to v1.x and v0.38.x

#### PR checklist

- [x] Tests written/updated
- [x] Changelog entry added in `.changelog` (we use
[unclog](https://github.com/informalsystems/unclog) to manage our
changelog)
- [x] Updated relevant documentation (`docs/` or `spec/`) and code
comments

(cherry picked from commit a7fd3a994d859c9cf7f451d0147cace816c94378)

# Conflicts:
#	.github/workflows/testapp-docker.yml