refactor: cometbft image build (#5179)
---

This is a refactor of the CometBFT docker image build.

The current state of this build is completely broken. It builds an amd64
and an arm64 version, and publishes each of them individually without
doing a multi-arch image. The result is only the arch that finishes last
ends up tagged.

This PR will make the single-arch builds tagged with arch specific
labels, and the current labels with be multi-arch builds.

#### PR checklist

- [x] Tests written/updated
- [x] Changelog entry added in `.changelog` (we use
[unclog](https://github.com/informalsystems/unclog) to manage our
changelog)
- [x] Updated relevant documentation (`docs/` or `spec/`) and code
comments

(cherry picked from commit fa76e354b1604d810e06381a78eacddd4f129024)

# Conflicts:
#	.github/workflows/cometbft-docker.yml