chore: reindex to event sink from cli (#5209)
This adds a reindex command to the CLI interface. The command was
already implemented, and included in a previous tendermint release
(`0.34.x`) but the backport was missed during the tendermint -> cometbft
transition. This is a useful command to repipe events to a destination
sink, for example because of a migration, or if an event was lost
because the sink was offline.

This has been a recurring source of pain for us (Penumbra), and we
didn't realize it existed until now. First, adding to `main`, and I will
open backports to `0.38.x` and `0.37.x`.

---

#### PR checklist
- [ ] Tests written/updated
- [ ] Changelog entry added in `.changelog` (we use
[unclog](https://github.com/informalsystems/unclog) to manage our
changelog)
- [ ] Updated relevant documentation (`docs/` or `spec/`) and code
comments

Co-authored-by: Zachary Becker <zachary@interchainlabs.io>
(cherry picked from commit c653eee08d7f2c44bd87fcca0b8a16666cd08fee)