chore: reset branch to v0.38.x (#5326)
Our main thinking in this decision is that we want to get all of the
features from v1 and v2 that were useful to folks (performance, peering,
BLS sigs, etc) without introducing the large set of API breakages and
risks that I'm sure your team is familiar with given your upgrade.
We'll be sharing a more detailed roadmap for the stack soon, but driving
cometBFT performance will be one of our main priorities.
The reality is that v1 has also been released officially and adopted by
some folks, so we will be sure to continue to provide support or smooth
upgrade paths

we have https://github.com/cometbft/cometbft/tree/main-backup to cover
all of the previous work that has been done in v1 and v2.

We will be selecting work from this branch to backport, but in order to
maintain stability, we will not be introducing a lot of the large-scale
refactors that have been done

---------

Co-authored-by: Abdul Malek <me@almk.dev>