chore: misleading error return in boundary check (#5400)
noticed that the function was returning `(false, err)` even though `err`
is always `nil` at this point. this could give the impression that an
error was possible here when it actually isn’t.

updated the code to return `(false, nil)` instead, which makes it clear
that hitting the boundary is not an error case.


#### PR checklist

- [ ] Tests written/updated
- [ ] Changelog entry added in `.changelog` (we use
[unclog](https://github.com/informalsystems/unclog) to manage our
changelog)
- [ ] Updated relevant documentation (`docs/` or `spec/`) and code
comments

Co-authored-by: Matt Acciai <matt.acciai@gmail.com>