chore(backport): remove update consensus loop routine (#5411)
---

Backports 3 previous PRs to main. I was hoping to only do
https://github.com/cometbft/cometbft/pull/5411/commits/3933a2076c0f35d4eff8aab5501a9efc79496002,
however this PR had dependencies on a previous large refactor of the
consensus reactors gossip routines. That refactor took place across
https://github.com/cometbft/cometbft/pull/5411/commits/6fd6aa76b41647ed5c21aa325208936a6d37ecb8
and
https://github.com/cometbft/cometbft/pull/5411/commits/5f1d4ed510422f78ae87d46d9c0685661065b2b8.
https://github.com/cometbft/cometbft/pull/5411/commits/6fd6aa76b41647ed5c21aa325208936a6d37ecb8
was originally breaking since it removed `*PeerState.PickSendVote()`, I
added it back in to avoid the break in
https://github.com/cometbft/cometbft/pull/5411/commits/4bbd3c73e320ba5a966a632033f77a19a0fe9954.
Additionally the main PR
https://github.com/cometbft/cometbft/pull/5411/commits/3933a2076c0f35d4eff8aab5501a9efc79496002
was breaking due to changing `ConsensusState.GetRoundState()` to not
return a pointer anymore, I modified it to return a pointer again to
avoid the break in
https://github.com/cometbft/cometbft/pull/5411/commits/910bd20480d6b12cf25d67b5159a72ec2de0210c.

When profiling consensus with Go's `FlightRecorder`, this goroutine
constantly being scheduled was polluting the traces and making it
difficult to read (this was also another change that I originally wanted
to backport with the perf changes, but it was too breaking to do so
then).

#### PR checklist

- [ ] Tests written/updated
- [ ] Changelog entry added in `.changelog` (we use
[unclog](https://github.com/informalsystems/unclog) to manage our
changelog)
- [ ] Updated relevant documentation (`docs/` or `spec/`) and code
comments

---------

Co-authored-by: Sergio Mena <sergio@informal.systems>
Co-authored-by: Daniel <daniel.cason@informal.systems>
Co-authored-by: Ethan Buchman <ethan@coinculture.info>
Co-authored-by: wesl-ee <w@wesl.ee>